Moved files, created folders
Created folders for RequestModels and BaseModels, updated namespaces.
