+++ PoohAPI/Models/Company.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using PoohAPI.Models.BaseModels;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
