+++ PoohAPI/Models/RequestModels/RegisterRequest.cs
@@ -4,7 +4,7 @@ using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace PoohAPI.Models
+namespace PoohAPI.Models.RequestModels
 {
     public class RegisterRequest
     {
